Major refactor

+++ src/bin/updater.js
@@ -1,27 +1,26 @@
-const courseScraper = require('../utils/courseScraper');
-const config = require('../config');
-const Course = require('../models/Course');
+const { scrapeCourses } = require('../services/scraper.service');
+const config = require('../config/config');
+const { Course } = require('../models');
+const logger = require('../config/logger');
 
-async function getCourseCodes () {
+async function updateCourseCodes() {
   try {
-    const collegeCourses = (await Promise.all(
-      config.COLLEGE_URLS.map((url, index) => courseScraper(index)),
-    )).reduce((courses, results) => courses.concat(results), []);
+    const collegeCourses = (
+      await Promise.all(config.COLLEGE_URLS.map((url, index) => scrapeCourses(index)))
+    ).reduce((courses, results) => courses.concat(results), []);
 
+    if (!collegeCourses || !collegeCourses.length) {
+      logger.error('No courses found');
+      return;
+    }
     await Course.deleteMany({});
     await Course.insertMany(collegeCourses);
+    logger.info('Updated course codes');
   } catch (error) {
-    console.error(error);
+    logger.error(error);
   }
 }
 
-module.exports = async () => {
-  setTimeout(async function updateCoursesLoop () {
-    try {
-      await getCourseCodes();
-      setTimeout(updateCoursesLoop, config.COURSE_UPDATE_INTERVAL);
-    } catch (error) {
-      console.log(error);
-    }
-  }, 0);
+module.exports = {
+  updateCourseCodes,
 };
